% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_transformations.R
\name{blue_equate}
\alias{blue_equate}
\title{This functions calculates parameters to equate measurements in a IRT framework}
\usage{
blue_equate(x, y, x_ee, y_ee)
}
\arguments{
\item{x}{vector of continous variable that you want to equate}

\item{y}{vector of continous variable that you want to equate to. This variable will have the same metric as parameter x.}

\item{x_ee}{vector of continous variable with estimated error.}

\item{y_ee}{vector of continous variable with estimated error.}
}
\value{
The output is a list of tibbles with the calculated parameters, the items and the difficulty of items of one group and the other. Finally a the plot with control bands to check item displacement.
}
\description{
It calculates mean-mean parameters in order to do a linear transformation of a scale to make it comparable with another one.
The function does the mean-mean transformation for test in a IRT framework, a good reference is:
Kolen, M & Brennan, R. (2014) Test equating, scaling and linking. Methods and practice. Third edition. NY: Springer.
}
\examples{
x1 <- c(1,2,3,4,5,6,7,8,9,10)
y1 <- c(2,3,4,5,6,7,9,8,6,5)
x.ee1 <- c(1,2,3,5,6,7,8,9,20,10)
y.ee1 <- c(9,8,7,6,5,4,3,2,1,2)
equate_table <- data.frame(x1,x.ee1,y1,y.ee1)
blue_equate(equate_table$x1,equate_table$y1,
equate_table$x.ee1,
equate_table$y.ee1)
}
\author{
Juan Carlos Saravia
}
