% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_transformations.R
\name{bluebare}
\alias{bluebare}
\title{This function can be used to do linear percentile scaling. It helps to create thresholds for surveys and continous variables.}
\arguments{
\item{x}{vector of continous values}

\item{sdev}{standard deviation which is to be assigned for the transformation, default value is 100.}

\item{means}{mean which is to be assigned for the linear transformation. Default value is 500.}

\item{threshold}{logical parameter were TRUE assigns three thresholds. It is based on a normal distribution were values lower than -1 standard deviation from the mean are low, values higher than 1 standard deviation are high and values in between are medium (default).
if logical parameter is FALSE, 5 thresholds are assigned where values bigger than 2 standard deviations are very high, higher than 1 standard deviation are high, higher than -1 are medium, higher than -2 are low and lower than -2 standard deviations are very low.}
}
\value{
The output is a tibble with raw scores, zscores and thresholds
}
\description{
It calculates the cumulative frequency of a continous variable, transforms this variable into percentiles by using a constant of .50 in order to calculate values over the 50% and under it based on the frequency of values.
After that, it standardize scores and then transforms values with a mean of 500 and standard deviation of 100 (default).
}
\examples{
\donttest{bluebare(x,sdev,means,threshold)}
}
\author{
Juan Carlos Saravia
}
