% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_validity_measures.R
\name{comp_reliability}
\alias{comp_reliability}
\title{Function to calculate Composite reliability}
\usage{
comp_reliability(x)
}
\arguments{
\item{x}{lavaan object: The name of the model fit that was calculated from the specified CFA model with the lavaan package.}
}
\value{
The output is a data.frame specifying the latent factor in the first column and the CR in the second column.
}
\description{
This function needs a lavaan object with a model fit to work. It calculates Composite reliability for CFA models.
There are two formulas inside the function and they run depending on the fit characteristics.
The composite reliability or factor rho coefficient is the ratio of explained variance over total variance.
With no error correlations the CR is just calculated as previously defined.
A different formula is needed for when indicators share at least one error covariance.
In this case the total variance is calculated by adding the sum of the unstandarized error variance multiplied by 2.
}
\examples{
#'set.seed(123456)
data <- data.frame(replicate(10,sample(1:5,1000,rep=TRUE)))

m2 <- 'f=~ X1 + X2 + X3+X4+X5+X6+X7+X8+X9+X10'
fit2 <- lavaan::cfa(m2, data = data)
lavaan::summary(fit2, fit.measures = TRUE,
       standardized = TRUE)
comp_reliability(fit2)

}
\seealso{
Kline, R. (2016). Principles and Practice of Structural Equation Modeling. Fourth Edition. Guilford press. NY.

Raykov, T. (2004). Behavioral scale realiability and measurement invariance evaluation using latent variable modeling. Behavior therapy, 35, 299-331.
}
\author{
Juan Carlos Saravia
}
