% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tables.R
\name{grafi}
\alias{grafi}
\title{Table of overall fit measures for a CFA and SEM models}
\arguments{
\item{x}{lavaan object: The name of the model fit that was calculated from the specified CFA model in the lavaan package.}

\item{y}{specify the amount of rows to show in the modification indexes table.}
}
\value{
The output is a list of tibbles that gives you overall fit indices, parameter estimates and modification indices for CFA models.
Example of fit indices: Chi square, degrees of freedom, p-values of Chi square, CFI, TLI, RMSEA, SRMR.
Example of parameter estimates: undstandardized betas, standardized betas, standard error, z values and p values
}
\description{
This function needs a lavaan object with a model fit to work. It creates a table of all fit indices for a CFA and SEM model.
}
\examples{
\donttest{grafi(fit)}
}
\author{
Juan Carlos Saravia
}
