% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tables.R
\name{grafi_distribution}
\alias{grafi_distribution}
\title{Table of distributions of categories in a vector}
\usage{
grafi_distribution(x)
}
\arguments{
\item{x}{vector of values that show a table of frequencies, proportions and cumulative frequencies.}
}
\value{
a tibble of 4 columns with categories, frequencies, proportions and cumulative frequencies.
}
\description{
This function just needs a vector and it is useful if you want to make a tidy table of distributions of items. It works with the map function from the purrr package which enables you to calculate multiple ordered tables of items.
}
\examples{
data_prueba <- c(1,2,3,4,5,6,7,78,7,7,7,7,7,7,7,7,8,8,8,8,8)
grafi_distribution(data_prueba)
}
\author{
Juan Carlos Saravia
}
