% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tables.R
\name{grafi_fit}
\alias{grafi_fit}
\title{Table of overall fit measures for a CFA and SEM models}
\usage{
grafi_fit(x, mi.nrows = 5)
}
\arguments{
\item{x}{lavaan object: The name of the model fit that was calculated from the specified CFA or SEM model in the lavaan package.}

\item{mi.nrows}{specify the amount of rows to show in the modification indexes table. The default is 5 rows.}
}
\value{
The output is a list of tibbles that gives you overall fit indices, parameter estimates and modification indices for CFA or SEM models.
Example of fit indices: Chi square, degrees of freedom, p-values of Chi square, CFI, TLI, RMSEA, SRMR.
Example of parameter estimates: undstandardized betas, standardized betas, standard error, z values and p values.
I want to thank Rose Hartmann because with her web page I was able to make one chunk of this function.
}
\description{
This function needs a lavaan object with a model fit to work. It creates a table of all fit indices for a SEM model.
}
\examples{
set.seed(123456)
data <- data.frame(replicate(10,sample(1:5,1000,rep=TRUE)))

m2 <- 'f=~ X1 + X2 + X3+X4+X5+X6+X7+X8+X9+X10'
fit2 <- lavaan::cfa(m2, data = data)
lavaan::summary(fit2, fit.measures = TRUE,
       standardized = TRUE)
grafi_fit(fit2)
}
\author{
Juan Carlos Saravia
}
