% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tables.R
\name{grafi_wrightmap}
\alias{grafi_wrightmap}
\title{Wright Map for items difficulty and person ability}
\usage{
grafi_wrightmap(persons.measure, items.measure, items.names, groups = 5)
}
\arguments{
\item{persons.measure}{Person's abilities}

\item{items.measure}{Item's difficulties}

\item{items.names}{Item's names}

\item{groups}{amounts of groups that the items are seperated}
}
\value{
a ggplot histogram of amount of people and a grapha of item difficulty
}
\description{
This function needs two vectors (person's abilities and items difficulties). With that information it will make a nice wright map
}
\examples{
/donttest{
personas.data <- data.frame(personas=sort(round(rnorm(500, 0,1),3)))

items.data <- data.frame(
  items_n = as.numeric(seq(1:36)),
  items_g = ceiling(runif(36, 0, 5)),
  items = sort(round(runif(36, -3,3),3)),
  stringsAsFactors=F)
items.data <- cbind(items.data,
                    items_c = dplyr::if_else(items.data$items_n < 10,
                     paste("IT_EX",items.data$items_n, sep="_0"),
                                     paste("IT_EX",items.data$items_n, sep="_")))


grafi_wrightmap(personas.data,items.data$items,items.data$items_c)}


}
\author{
Juan Carlos Saravia
}
