% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_validity_measures.R
\name{avar_extracted}
\alias{avar_extracted}
\title{Function to calculate Average variance extracted (AVE)}
\usage{
avar_extracted(x)
}
\arguments{
\item{x}{lavaan object: The name of the model fit that was calculated from the specified CFA model in the lavaan package.}
}
\value{
The output is a data.frame specifying the latent factor in the first column and the AVE in the second column.
}
\description{
This function needs a lavaan object with a model fit to work. It calculates the Average variance extracted for CFA models.
It is the average of the squared standardized pattern coefficients for indicators that depend on the same factor but are specified to measure no other factors.
}
\examples{
#' #'set.seed(123456)
data <- data.frame(replicate(10,sample(1:5,1000,rep=TRUE)))

m2 <- 'f=~ X1 + X2 + X3+X4+X5+X6+X7+X8+X9+X10'
fit2 <- lavaan::cfa(m2, data = data)
lavaan::summary(fit2, fit.measures = TRUE,
       standardized = TRUE)
avar_extracted(fit2)


}
\seealso{
Kline, R. (2016). Principles and Practice of Structural Equation Modeling. Fourth Edition. Guilford press. NY.

Raykov, T. (2004). Behavioral scale realiability and measurement invariance evaluation using latent variable modeling. Behavior therapy, 35, 299-331.
}
\author{
Juan Carlos Saravia
}
